feat: 바로결제 쿠폰 적용시 최소주문 금액 확인 및 테스트 추가
