+++ backend/main-service/src/main/java/kkakka/mainservice/common/exception/OutOfMinOrderPriceException.java
@@ -1,0 +1,5 @@
+package kkakka.mainservice.common.exception;
+
+public class OutOfMinOrderPriceException extends KkaKkaException{
+
+}
