+++ backend/main-service/src/main/java/kkakka/mainservice/order/application/OrderService.java
@@ -2,12 +2,14 @@ package kkakka.mainservice.order.application;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import kkakka.mainservice.common.exception.KkaKkaException;
 import kkakka.mainservice.common.exception.NotFoundCouponException;
 import kkakka.mainservice.common.exception.NotFoundMemberException;
 import kkakka.mainservice.common.exception.NotFoundProductException;
 import kkakka.mainservice.common.exception.NotOrderOwnerException;
+import kkakka.mainservice.common.exception.OutOfMinOrderPriceException;
 import kkakka.mainservice.coupon.domain.Coupon;
 import kkakka.mainservice.coupon.domain.MemberCoupon;
 import kkakka.mainservice.coupon.domain.repository.CouponRepository;
@@ -68,11 +70,13 @@ public class OrderService {
             Product product = productRepository.findById(productId)
                 .orElseThrow(NotFoundProductException::new);
             ProductOrder productOrder = ProductOrder.create(product, product.getPrice(), quantity);
-            if (couponId != null) {
+            if (!Objects.isNull(couponId)) {
                 Coupon coupon = couponRepository.findById(couponId).orElseThrow(NotFoundCouponException::new);
+                if (product.getDiscountPrice() * quantity < coupon.getMinOrderPrice()) {
+                    throw new OutOfMinOrderPriceException();
+                }
                 productOrder.applyCoupon(coupon);
             }
-
             productOrders.add(productOrder);
             orderTotalPrice += productOrder.getTotalPrice();
         }
