+++ backend/main-service/src/test/java/kkakka/mainservice/coupon/acceptance/CouponAcceptanceTest.java
@@ -12,16 +12,12 @@ import io.restassured.response.Response;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.util.ArrayList;
-import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import kkakka.mainservice.DocumentConfiguration;
 import kkakka.mainservice.member.auth.ui.dto.SocialProviderCodeRequest;
 import kkakka.mainservice.member.member.domain.ProviderName;
 import kkakka.mainservice.order.application.dto.ProductOrderDto;
-import kkakka.mainservice.order.ui.dto.OrderRequest;
-import kkakka.mainservice.order.ui.dto.RecipientRequest;
 import org.assertj.core.api.Assertions;
 import org.hibernate.Session;
 import org.junit.jupiter.api.AfterEach;
@@ -451,35 +447,6 @@ public class CouponAcceptanceTest extends DocumentConfiguration {
 
         // then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
-
-    @DisplayName("쿠폰 적용 결제 - 성공")
-    @Test
-    public void orderWithCoupon() {
-        // given
-        tearDown();
-        String accessToken = 액세스_토큰_가져옴();
-        String couponId = 상품_쿠폰_다운로드(accessToken);
-        ProductOrderDto productOrderDto = new ProductOrderDto(PRODUCT_1.getId(), Long.parseLong(couponId), 3);
-        List<ProductOrderDto> productOrderDtoList = new ArrayList<>();
-        productOrderDtoList.add(productOrderDto);
-        OrderRequest orderRequest = new OrderRequest(
-            new RecipientRequest(TEST_MEMBER_01.getName(), TEST_MEMBER_01.getEmail(),
-            TEST_MEMBER_01.getPhone(), TEST_MEMBER_01.getAddress()), productOrderDtoList);
-
-        // when
-        final ExtractableResponse<Response> response = RestAssured
-            .given(spec).log().all()
-            .header("Authorization", "Bearer " + accessToken)
-            .filter(document("order-with-coupon"))
-            .contentType(MediaType.APPLICATION_JSON_VALUE)
-            .body(orderRequest)
-            .when()
-            .post("/api/orders/")
-            .then().log().all().extract();
-
-        // then
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
     }
 
     @DisplayName("상품 바로주문 쿠폰 적용 취소 - 성공")
